tracker improvements
