+++ src/main/java/task/client/TorrentClientConsole.java
@@ -1,0 +1,9 @@
+package task.client;
+
+/**
+ * Created by equi on 14.03.16.
+ *
+ * @author Kravchenko Dima
+ */
+public class TorrentClientConsole {
+}
